python/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2
argparse==1.4.0